requirements.txt
asgiref
djangorestframework
django
django-cors-headers
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
